priceの小数点対応

+++ api/db/migrate/20230119002326_create_products.rb
@@ -29,10 +29,10 @@ class CreateProducts < ActiveRecord::Migration[7.0]
       t.integer :weight
       t.integer :img_count
       t.integer :price
-      t.integer :discogs_price
+      t.float :discogs_price
       t.integer :discogs_release_id
       t.bigint :discogs_listing_id
-      t.decimal :ebay_price
+      t.float :ebay_price
       t.integer :ebay_id
       t.integer :master_id
       t.string :cost_price
